+++ test/flycheck-test.el
@@ -1586,11 +1586,13 @@ and extension, as in `file-name-base'."
          :checker python-flake8)
      '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
          :checker python-flake8)
+     '(9 9 info "function name should be lowercase"
+         :checker python-flake8 :id "N802")
      '(12 29 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
      '(12 31 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
-     '(22 1 warning "undefined name 'antigravity'" :id "F821"
+     '(22 1 error "undefined name 'antigravity'" :id "F821"
           :checker python-flake8))
     ;; Selecting Pylint should give us its errors
     (flycheck-select-checker 'python-pylint)
@@ -1667,11 +1669,13 @@ and extension, as in `file-name-base'."
          :checker python-flake8)
      '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
          :checker python-flake8)
+     '(9 9 info "function name should be lowercase"
+         :checker python-flake8 :id "N802")
      '(12 29 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
      '(12 31 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
-     '(22 1 warning "undefined name 'antigravity'" :id "F821"
+     '(22 1 error "undefined name 'antigravity'" :id "F821"
           :checker python-flake8))))
 
 (ert-deftest flycheck/selects-checker-automatically/first-enabled-checker ()
@@ -1689,11 +1693,13 @@ and extension, as in `file-name-base'."
          :checker python-flake8)
      '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
          :checker python-flake8)
+     '(9 9 info "function name should be lowercase"
+         :checker python-flake8 :id "N802")
      '(12 29 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
      '(12 31 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
-     '(22 1 warning "undefined name 'antigravity'" :id "F821"
+     '(22 1 error "undefined name 'antigravity'" :id "F821"
           :checker python-flake8))))
 
 (ert-deftest flycheck/selects-checker-automatically/no-disabled-checker ()
@@ -4462,7 +4468,7 @@ Why not:
   (let ((python-indent-guess-indent-offset nil))       ; Silence Python Mode!
     (flycheck-ert-should-syntax-check
      "checkers/python-syntax-error.py" 'python-mode
-     '(3 13 warning "SyntaxError: invalid syntax" :id "E901"
+     '(3 13 error "SyntaxError: invalid syntax" :id "E901"
          :checker python-flake8))))
 
 (flycheck-ert-def-checker-test python-flake8 python warning-ignored
@@ -4471,7 +4477,9 @@ Why not:
      "checkers/python/test.py" 'python-mode
      '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
          :checker python-flake8)
-     '(22 1 warning "undefined name 'antigravity'" :id "F821"
+     '(9 9 info "function name should be lowercase"
+         :checker python-flake8 :id "N802")
+     '(22 1 error "undefined name 'antigravity'" :id "F821"
           :checker python-flake8))))
 
 (flycheck-ert-def-checker-test python-flake8 python maximum-complexity
@@ -4482,13 +4490,15 @@ Why not:
          :checker python-flake8)
      '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
          :checker python-flake8)
+     '(9 9 info "function name should be lowercase"
+         :checker python-flake8 :id "N802")
      '(12 1 warning "'Spam.with_ham' is too complex (3)" :id "C901"
           :checker python-flake8)
      '(12 29 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
      '(12 31 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
-     '(22 1 warning "undefined name 'antigravity'" :id "F821"
+     '(22 1 error "undefined name 'antigravity'" :id "F821"
           :checker python-flake8))))
 
 (flycheck-ert-def-checker-test python-flake8 python error-maximum-line-length
@@ -4499,13 +4509,15 @@ Why not:
          :checker python-flake8)
      '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
          :checker python-flake8)
+     '(9 9 info "function name should be lowercase"
+         :checker python-flake8 :id "N802")
      '(10 46 warning "line too long (46 > 45 characters)" :id "E501"
           :checker python-flake8)
      '(12 29 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
      '(12 31 warning "unexpected spaces around keyword / parameter equals"
           :id "E251" :checker python-flake8)
-     '(22 1 warning "undefined name 'antigravity'" :id "F821"
+     '(22 1 error "undefined name 'antigravity'" :id "F821"
           :checker python-flake8))))
 
 (flycheck-ert-def-checker-test python-flake8 python nil
@@ -4515,11 +4527,13 @@ Why not:
        :checker python-flake8)
    '(7 1 warning "expected 2 blank lines, found 1" :id "E302"
        :checker python-flake8)
+   '(9 9 info "function name should be lowercase"
+       :checker python-flake8 :id "N802")
    '(12 29 warning "unexpected spaces around keyword / parameter equals"
         :id "E251" :checker python-flake8)
    '(12 31 warning "unexpected spaces around keyword / parameter equals"
         :id "E251" :checker python-flake8)
-   '(22 1 warning "undefined name 'antigravity'" :id "F821"
+   '(22 1 error "undefined name 'antigravity'" :id "F821"
         :checker python-flake8)))
 
 (flycheck-ert-def-checker-test python-pylint python syntax-error